+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Reworked network requests to resolve CORS issues in Firefox.
 - Added support for Automatic Persisted Queries (APQ) in the DevTools panel.
+- Added a divider to the request list in the DevTools panel when a navigation
+  occurs during Preserve Log mode.
 
 ## 0.2.0 (2026-03-22)